Remove explicit boxing and unboxing

+++ src/main/java/com/microsoft/sqlserver/jdbc/Column.java
@@ -191,7 +191,7 @@ final class Column {
     }
 
     int getInt(TDSReader tdsReader) throws SQLServerException {
-        return ((Integer) getValue(JDBCType.INTEGER, null, null, tdsReader)).intValue();
+        return (Integer) getValue(JDBCType.INTEGER, null, null, tdsReader);
     }
 
     void updateValue(JDBCType jdbcType,
