+++ src/main/java/com/microsoft/sqlserver/jdbc/FailOverInfo.java
@@ -71,7 +71,7 @@ final class FailoverInfo {
                 instancePort = con.getInstancePort(failoverPartner, instanceValue);
 
                 try {
-                    portNumber = (new Integer(instancePort)).intValue();
+                    portNumber = new Integer(instancePort);
                 }
                 catch (NumberFormatException e) {
                     // Should not get here as the server should give a proper port number anyway.
